test: add fixtures
add task data and test helpers (utils)
