+++ miamoremustachio/ToDo/tests/fixtures/test-utils.js
@@ -1,0 +1,7 @@
+function getFinalResourceId(endpoint) {
+  const startIndex = endpoint.lastIndexOf('/') + 1;
+  const resourceId = endpoint.slice(startIndex);
+  return resourceId;
+}
+
+module.exports = { getFinalResourceId };
